Deal with submodules.
