+++ docs/src/submodule.md
@@ -14,17 +14,27 @@ submodule generated by a finite list of elements in the original module.
 We define two submodules to be equal if they are (transitively) submodules
 of the same module $M$ and their generators generate the same set of elements.
 
-As well as implementing the entire Module interface, AbstractAlgebra
-submodules also provide the following interface.
+## Generic submodule type
+
+AbstractAlgebra implements a generic submodule type `Generic.Submodule{T}`
+where `T` is the element type of the base ring. See `src/generic/GenericTypes.jl`
+for more details.
+
+Elements of a generic submodule have type `Generic.SubmoduleElem{T}`.
+
+## Abstract types
+
+Submodule types belong to the abstract type `FPModule{T}` and their elements
+to `FPModuleElem{T}`.
 
 ## Constructors
 
 ```@docs
-sub(::AbstractAlgebra.FPModule{T}, ::Vector{AbstractAlgebra.FPModuleElem{T}}) where T <: RingElement
+sub(::FPModule{T}, ::Vector{FPModuleElem{T}}) where T <: RingElement
 ```
 
 ```@docs
-sub(::AbstractAlgebra.FPModule{T}, ::Vector{Generic.Submodule{T}}) where T <: RingElement
+sub(::FPModule{T}, ::Vector{Generic.Submodule{T}}) where T <: RingElement
 ```
 
 Note that the preimage of the canonical injection can be obtained using the
@@ -82,19 +92,19 @@ following functionality.
 ### Basic manipulation
 
 ```@docs
-supermodule(::AbstractAlgebra.Generic.Submodule{T}) where T <: RingElement
+supermodule(::Generic.Submodule{T}) where T <: RingElement
 ```
 
 ```@docs
-issubmodule(::AbstractAlgebra.FPModule{T}, ::AbstractAlgebra.FPModule{T}) where T <: RingElement
+issubmodule(::FPModule{T}, ::FPModule{T}) where T <: RingElement
 ```
 
 ```@docs
-iscompatible(::AbstractAlgebra.FPModule{T}, ::AbstractAlgebra.FPModule{T}) where T <: RingElement
+iscompatible(::FPModule{T}, ::FPModule{T}) where T <: RingElement
 ```
 
 ```@docs
-dim(N::AbstractAlgebra.Generic.Submodule{T}) where T <: FieldElement
+dim(N::Generic.Submodule{T}) where T <: FieldElement
 ```
 
 **Examples**
@@ -157,7 +167,7 @@ julia> dim(N)
 ### Intersection
 
 ```@docs
-Base.intersect(M::AbstractAlgebra.FPModule{T}, N::AbstractAlgebra.FPModule{T}) where
+Base.intersect(M::FPModule{T}, N::FPModule{T}) where
 T <: RingElement
 ```
