+++ src/AbstractAlgebra.jl
@@ -415,6 +415,7 @@ include("AbsSeries.jl")
 include("RelSeries.jl")
 include("LaurentPoly.jl")
 include("FreeModule.jl")
+include("Submodule.jl")
 
 ###############################################################################
 #
@@ -474,7 +475,7 @@ import .Generic: abs_series, abs_series_type, add!, addeq!,
                  right_kernel, section_map, setcoeff!,
                  set_exponent_vector!, set_field!, set_limit!,
                  setpermstyle, size,
-                 sort_terms!, sub, summands,
+                 sort_terms!, summands,
                  supermodule, term, terms, total_degree,
                  to_univariate, trailing_coefficient,
 		 truncate, upscale, var_index, vars,
@@ -543,7 +544,7 @@ export abs_series, abs_series_type, add!, addeq!,
                  section_map, setcoeff!,
                  set_exponent_vector!, set_field!, set_limit!,
                  setpermstyle,
-                 size, sort_terms!, sub, subst, summands, supermodule,
+                 size, sort_terms!, subst, summands, supermodule,
                  sylvester_matrix, term, terms, to_univariate,
                  total_degree, trailing_coefficient, truncate,
                  upscale, var_index, vars, zero,
@@ -736,35 +737,6 @@ end
 
 function FunctionField(p::Generic.Poly{Generic.Rat{T}}, s::AbstractString; cached::Bool=true) where T <: FieldElement
    Generic.FunctionField(p, s; cached=cached)
-end
-
-@doc Markdown.doc"""
-    VectorSpace(R::Field, dim::Int; cached::Bool = true)
-
-Return the vector space over the field $R$ with the given dimension.
-"""
-function VectorSpace(R::Field, dim::Int; cached::Bool = true)
-   Generic.FreeModule(R, dim; cached=cached)
-end
-
-function vector_space(R::Field, dim::Int; cached::Bool = true)
-   Generic.FreeModule(R, dim; cached=cached)
-end
-
-@doc Markdown.doc"""
-    sub(m::Module{T}, gens::Vector{<:ModuleElem{T}}) where T <: RingElement
-
-Return the submodule `S` of the module `m` generated by the given generators,
-given as elements of `m`, and a map which is the canonical injection from `S`
-to `m`.
-"""
-function sub(m::Module{T}, gens::Vector{<:ModuleElem{T}}) where T <: RingElement
-   Generic.sub(m, gens)
-end
-
-# Handles empty vector of elements
-function sub(m::Module{T}, gens::Vector{Any}) where T <: RingElement
-   Generic.sub(m, gens)
 end
 
 @doc Markdown.doc"""
