+++ src/generic/Submodule.jl
@@ -4,8 +4,6 @@
 #
 ###############################################################################
 
-export Submodule, SubmoduleElem, ngens, gens, supermodule, sub
-
 ###############################################################################
 #
 #   Basic manipulation
@@ -32,11 +30,11 @@ function gen(N::Submodule{T}, i::Int) where T <: RingElement
 end
 
 @doc Markdown.doc"""
-    dim(N::AbstractAlgebra.Generic.Submodule{T}) where T <: FieldElement
+    dim(N::Submodule{T}) where T <: FieldElement
 
 Return the dimension of the given vector subspace.
 """
-dim(N::AbstractAlgebra.Generic.Submodule{T}) where T <: FieldElement = length(N.gen_cols)
+dim(N::Submodule{T}) where T <: FieldElement = length(N.gen_cols)
 
 # Generators as elements of supermodule. Used internally.
 generators(N::Submodule{T}) where T <: RingElement = N.gens::Vector{elem_type(N.m)}
@@ -125,23 +123,6 @@ end
 #
 ###############################################################################
 
-reduced_form(mat::AbstractAlgebra.MatElem{T}) where T <: RingElement = hnf(mat)
-
-function reduced_form(mat::AbstractAlgebra.MatElem{T}) where T <: FieldElement
-  r, m = rref(mat)
-  return m
-end
-
-isreduced_form(mat::AbstractAlgebra.MatElem{T}) where T <: RingElement = ishnf(mat)
-
-isreduced_form(mat::AbstractAlgebra.MatElem{T}) where T <: FieldElement = isrref(mat)
-
-@doc Markdown.doc"""
-    sub(m::AbstractAlgebra.FPModule{T}, gens::Vector{<:AbstractAlgebra.FPModuleElem{T}}) where T <: RingElement
-
-Return the submodule of the module `m` generated by the given generators,
-given as elements of `m`.
-"""
 function sub(m::AbstractAlgebra.FPModule{T}, gens::Vector{S}) where {T <: RingElement, S <: AbstractAlgebra.FPModuleElem{T}}
    R = base_ring(m)
    r = length(gens)
